+++ app/src/main/java/com/example/app21try6/database/models/BrandProductModel.kt
@@ -3,6 +3,6 @@ package com.example.app21try6.database.models
 data class BrandProductModel (
     var id:Int,
     var brandId:Long?,
-    var parentId:Int?,
+    var parentId:Long?,
     var name:String
 )
